refactoring of CSVReader.dump: define private class method .array_to_csv
I do not adopt "records.map {|rec| rec.to_csv }.join",
because I'm not sure if it is sufficient for certain data or not.
For example, a field value may contain carriage returns or line feeds,
and in that case, improper handling of the end of record would be damaging.